python-scripts/requirements.txt
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
PyPDF2==3.0.1
pikepdf==9.4.2
reportlab==4.2.5
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
pdf2docx==0.5.8
opencv-python==4.10.0.84
numpy==2.1.3
PyMuPDF
